+++ backend/llm/qa_base.py
@@ -15,21 +15,17 @@ from langchain.prompts.chat import (
     HumanMessagePromptTemplate,
     SystemMessagePromptTemplate,
 )
+from llm.utils.format_chat_history import format_chat_history
 from llm.utils.get_prompt_to_use import get_prompt_to_use
 from llm.utils.get_prompt_to_use_id import get_prompt_to_use_id
 from logger import get_logger
 from models import BrainSettings  # Importing settings related to the 'brain'
-from models.chats import ChatQuestion
-from models.databases.supabase.chats import CreateChatHistory
 from modules.brain.service.brain_service import BrainService
+from modules.chat.dto.chats import ChatQuestion
+from modules.chat.dto.inputs import CreateChatHistory
+from modules.chat.dto.outputs import GetChatHistoryOutput
+from modules.chat.service.chat_service import ChatService
 from pydantic import BaseModel
-from repository.chat import (
-    GetChatHistoryOutput,
-    format_chat_history,
-    get_chat_history,
-    update_chat_history,
-    update_message_by_id,
-)
 from supabase.client import Client, create_client
 from vectorstore.supabase import CustomSupabaseVectorStore
 
@@ -40,6 +36,7 @@ QUIVR_DEFAULT_PROMPT = "Your name is Quivr. You're a helpful assistant.  If you 
 
 
 brain_service = BrainService()
+chat_service = ChatService()
 
 
 class QABaseBrainPicking(BaseModel):
@@ -92,7 +89,7 @@ class QABaseBrainPicking(BaseModel):
             return OllamaEmbeddings(
                 base_url=self.brain_settings.ollama_api_base_url
             )  # pyright: ignore reportPrivateUsage=none
-        else: 
+        else:
             return OpenAIEmbeddings()
 
     supabase_client: Optional[Client] = None
@@ -155,7 +152,6 @@ class QABaseBrainPicking(BaseModel):
         if self.brain_settings.ollama_api_base_url and model.startswith("ollama"):
             api_base = self.brain_settings.ollama_api_base_url
 
-
         return ChatLiteLLM(
             temperature=temperature,
             max_tokens=self.max_tokens,
@@ -163,7 +159,7 @@ class QABaseBrainPicking(BaseModel):
             streaming=streaming,
             verbose=False,
             callbacks=callbacks,
-            api_base= api_base
+            api_base=api_base,
         )  # pyright: ignore reportPrivateUsage=none
 
     def _create_prompt_template(self):
@@ -192,7 +188,9 @@ class QABaseBrainPicking(BaseModel):
     def generate_answer(
         self, chat_id: UUID, question: ChatQuestion
     ) -> GetChatHistoryOutput:
-        transformed_history = format_chat_history(get_chat_history(self.chat_id))
+        transformed_history = format_chat_history(
+            chat_service.get_chat_history(self.chat_id)
+        )
         answering_llm = self._create_llm(
             model=self.model,
             streaming=False,
@@ -230,7 +228,7 @@ class QABaseBrainPicking(BaseModel):
 
         answer = model_response["answer"]
 
-        new_chat = update_chat_history(
+        new_chat = chat_service.update_chat_history(
             CreateChatHistory(
                 **{
                     "chat_id": chat_id,
@@ -264,7 +262,7 @@ class QABaseBrainPicking(BaseModel):
     async def generate_stream(
         self, chat_id: UUID, question: ChatQuestion
     ) -> AsyncIterable:
-        history = get_chat_history(self.chat_id)
+        history = chat_service.get_chat_history(self.chat_id)
         callback = AsyncIteratorCallbackHandler()
         self.callbacks = [callback]
 
@@ -323,7 +321,7 @@ class QABaseBrainPicking(BaseModel):
         if question.brain_id:
             brain = brain_service.get_brain_by_id(question.brain_id)
 
-        streamed_chat_history = update_chat_history(
+        streamed_chat_history = chat_service.update_chat_history(
             CreateChatHistory(
                 **{
                     "chat_id": chat_id,
@@ -386,7 +384,7 @@ class QABaseBrainPicking(BaseModel):
         assistant += sources_string
 
         try:
-            update_message_by_id(
+            chat_service.update_message_by_id(
                 message_id=str(streamed_chat_history.message_id),
                 user_message=question.question,
                 assistant=assistant,
