+++ backend/llm/qa_headless.py
@@ -7,31 +7,29 @@ from langchain.callbacks.streaming_aiter import AsyncIteratorCallbackHandler
 from langchain.chains import LLMChain
 from langchain.chat_models import ChatLiteLLM
 from langchain.chat_models.base import BaseChatModel
-from models import BrainSettings  # Importing settings related to the 'brain'
 from langchain.prompts.chat import ChatPromptTemplate, HumanMessagePromptTemplate
-from logger import get_logger
-from models.chats import ChatQuestion
-from models.databases.supabase.chats import CreateChatHistory
-from modules.prompt.entity.prompt import Prompt
-from pydantic import BaseModel
-from repository.chat import (
-    GetChatHistoryOutput,
+from llm.utils.format_chat_history import (
     format_chat_history,
     format_history_to_openai_mesages,
-    get_chat_history,
-    update_chat_history,
-    update_message_by_id,
 )
-
 from llm.utils.get_prompt_to_use import get_prompt_to_use
 from llm.utils.get_prompt_to_use_id import get_prompt_to_use_id
+from logger import get_logger
+from models import BrainSettings  # Importing settings related to the 'brain'
+from modules.chat.dto.chats import ChatQuestion
+from modules.chat.dto.inputs import CreateChatHistory
+from modules.chat.dto.outputs import GetChatHistoryOutput
+from modules.chat.service.chat_service import ChatService
+from modules.prompt.entity.prompt import Prompt
+from pydantic import BaseModel
 
 logger = get_logger(__name__)
 SYSTEM_MESSAGE = "Your name is Quivr. You're a helpful assistant. If you don't know the answer, just say that you don't know, don't try to make up an answer.When answering use markdown or any other techniques to display the content in a nice and aerated way."
+chat_service = ChatService()
 
 
 class HeadlessQA(BaseModel):
-    brain_settings = BrainSettings() 
+    brain_settings = BrainSettings()
     model: str
     temperature: float = 0.0
     max_tokens: int = 2000
@@ -104,7 +102,10 @@ class HeadlessQA(BaseModel):
     def generate_answer(
         self, chat_id: UUID, question: ChatQuestion
     ) -> GetChatHistoryOutput:
-        transformed_history = format_chat_history(get_chat_history(self.chat_id))
+        # Move format_chat_history to chat service ?
+        transformed_history = format_chat_history(
+            chat_service.get_chat_history(self.chat_id)
+        )
         prompt_content = (
             self.prompt_to_use.content if self.prompt_to_use else SYSTEM_MESSAGE
         )
@@ -120,7 +121,7 @@ class HeadlessQA(BaseModel):
         model_prediction = answering_llm.predict_messages(messages)
         answer = model_prediction.content
 
-        new_chat = update_chat_history(
+        new_chat = chat_service.update_chat_history(
             CreateChatHistory(
                 **{
                     "chat_id": chat_id,
@@ -152,7 +153,9 @@ class HeadlessQA(BaseModel):
         callback = AsyncIteratorCallbackHandler()
         self.callbacks = [callback]
 
-        transformed_history = format_chat_history(get_chat_history(self.chat_id))
+        transformed_history = format_chat_history(
+            chat_service.get_chat_history(self.chat_id)
+        )
         prompt_content = (
             self.prompt_to_use.content if self.prompt_to_use else SYSTEM_MESSAGE
         )
@@ -186,7 +189,7 @@ class HeadlessQA(BaseModel):
             ),
         )
 
-        streamed_chat_history = update_chat_history(
+        streamed_chat_history = chat_service.update_chat_history(
             CreateChatHistory(
                 **{
                     "chat_id": chat_id,
@@ -221,7 +224,7 @@ class HeadlessQA(BaseModel):
         await run
         assistant = "".join(response_tokens)
 
-        update_message_by_id(
+        chat_service.update_message_by_id(
             message_id=str(streamed_chat_history.message_id),
             user_message=question.question,
             assistant=assistant,
