+++ backend/models/__init__.py
@@ -1,18 +1,6 @@
 from .brains_subscription_invitations import BrainSubscription
-from .chat import Chat, ChatHistory
-from .chats import ChatMessage, ChatQuestion
 from .files import File
 from .settings import (BrainRateLimiting, BrainSettings, ResendSettings,
                        get_documents_vector_store, get_embeddings,
                        get_supabase_client, get_supabase_db)
 from .user_usage import UserUsage
-
-# TODO uncomment the below import when start using SQLalchemy
-# from .sqlalchemy_repository import (
-#     User,
-#     Brain,
-#     BrainUser,
-#     BrainVector,
-#     BrainSubscriptionInvitation,
-#     ApiKey
-# )
