+++ backend/models/databases/supabase/__init__.py
@@ -1,7 +1,6 @@
 from models.databases.supabase.api_brain_definition import ApiBrainDefinitions
 from models.databases.supabase.brains_subscription_invitations import \
     BrainSubscription
-from models.databases.supabase.chats import Chats
 from models.databases.supabase.files import File
 from models.databases.supabase.user_usage import UserUsage
 from models.databases.supabase.vectors import Vector
