+++ backend/models/databases/supabase/supabase.py
@@ -2,7 +2,6 @@ from logger import get_logger
 from models.databases.supabase import (
     ApiBrainDefinitions,
     BrainSubscription,
-    Chats,
     File,
     UserUsage,
     Vector,
@@ -15,7 +14,6 @@ class SupabaseDB(
     UserUsage,
     File,
     BrainSubscription,
-    Chats,
     Vector,
     ApiBrainDefinitions,
 ):
@@ -24,6 +22,5 @@ class SupabaseDB(
         UserUsage.__init__(self, supabase_client)
         File.__init__(self, supabase_client)
         BrainSubscription.__init__(self, supabase_client)
-        Chats.__init__(self, supabase_client)
         Vector.__init__(self, supabase_client)
         ApiBrainDefinitions.__init__(self, supabase_client)
