+++ backend/modules/chat/controller/__init__.py
@@ -1,0 +1,1 @@
+from .chat_routes import chat_router
