+++ backend/modules/chat/controller/chat_routes.py
@@ -7,36 +7,27 @@ from fastapi.responses import StreamingResponse
 from llm.qa_base import QABaseBrainPicking
 from llm.qa_headless import HeadlessQA
 from middlewares.auth import AuthBearer, get_current_user
-from models import Chat, ChatQuestion, UserUsage, get_supabase_db
-from models.databases.supabase.chats import QuestionAndAnswer
+from models.user_usage import UserUsage
 from modules.brain.service.brain_service import BrainService
-from modules.notification.service.notification_service import NotificationService
-from modules.user.entity.user_identity import UserIdentity
-from repository.chat import (
+from modules.chat.controller.chat.factory import get_chat_strategy
+from modules.chat.controller.chat.utils import NullableUUID, check_user_requests_limit
+from modules.chat.dto.chats import ChatItem, ChatQuestion
+from modules.chat.dto.inputs import (
     ChatUpdatableProperties,
     CreateChatProperties,
-    GetChatHistoryOutput,
-    create_chat,
-    get_chat_by_id,
-    get_user_chats,
-    update_chat,
+    QuestionAndAnswer,
 )
-from repository.chat.add_question_and_answer import add_question_and_answer
-from repository.chat.get_chat_history_with_notifications import (
-    ChatItem,
-    get_chat_history_with_notifications,
-)
-from routes.chat.factory import get_chat_strategy
-from routes.chat.utils import (
-    NullableUUID,
-    check_user_requests_limit,
-    delete_chat_from_db,
-)
+from modules.chat.dto.outputs import GetChatHistoryOutput
+from modules.chat.entity.chat import Chat
+from modules.chat.service.chat_service import ChatService
+from modules.notification.service.notification_service import NotificationService
+from modules.user.entity.user_identity import UserIdentity
 
 chat_router = APIRouter()
 
 notification_service = NotificationService()
 brain_service = BrainService()
+chat_service = ChatService()
 
 
 @chat_router.get("/chat/healthz", tags=["Health"])
@@ -56,7 +47,7 @@ async def get_chats(current_user: UserIdentity = Depends(get_current_user)):
     This endpoint retrieves all the chats associated with the current authenticated user. It returns a list of chat objects
     containing the chat ID and chat name for each chat.
     """
-    chats = get_user_chats(str(current_user.id))
+    chats = chat_service.get_user_chats(str(current_user.id))
     return {"chats": chats}
 
 
@@ -68,10 +59,9 @@ async def delete_chat(chat_id: UUID):
     """
     Delete a specific chat by chat ID.
     """
-    supabase_db = get_supabase_db()
     notification_service.remove_chat_notifications(chat_id)
 
-    delete_chat_from_db(supabase_db=supabase_db, chat_id=chat_id)
+    chat_service.delete_chat_from_db(chat_id)
     return {"message": f"{chat_id}  has been deleted."}
 
 
@@ -83,18 +73,20 @@ async def update_chat_metadata_handler(
     chat_data: ChatUpdatableProperties,
     chat_id: UUID,
     current_user: UserIdentity = Depends(get_current_user),
-) -> Chat:
+):
     """
     Update chat attributes
     """
 
-    chat = get_chat_by_id(chat_id)  # pyright: ignore reportPrivateUsage=none
+    chat = chat_service.get_chat_by_id(
+        chat_id  # pyright: ignore reportPrivateUsage=none
+    )
     if str(current_user.id) != chat.user_id:
         raise HTTPException(
             status_code=403,  # pyright: ignore reportPrivateUsage=none
             detail="You should be the owner of the chat to update it.",  # pyright: ignore reportPrivateUsage=none
         )
-    return update_chat(chat_id=chat_id, chat_data=chat_data)
+    return chat_service.update_chat(chat_id=chat_id, chat_data=chat_data)
 
 
 # create new chat
@@ -107,7 +99,7 @@ async def create_chat_handler(
     Create a new chat with initial chat messages.
     """
 
-    return create_chat(user_id=current_user.id, chat_data=chat_data)
+    return chat_service.create_chat(user_id=current_user.id, chat_data=chat_data)
 
 
 # add new question to chat
@@ -271,7 +263,7 @@ async def get_chat_history_handler(
     chat_id: UUID,
 ) -> List[ChatItem]:
     # TODO: RBAC with current_user
-    return get_chat_history_with_notifications(chat_id)
+    return chat_service.get_chat_history_with_notifications(chat_id)
 
 
 @chat_router.post(
@@ -286,4 +278,4 @@ async def add_question_and_answer_handler(
     """
     Add a new question and anwser to the chat.
     """
-    return add_question_and_answer(chat_id, question_and_answer)
+    return chat_service.add_question_and_answer(chat_id, question_and_answer)
