--- backend/repository/brain/delete_brain_secrets.py
@@ -1,3 +1,0 @@
-from modules.brain.service.brain_user_service import BrainUserService
-
-brain_user_service = BrainUserService()
